GHUB-0182, GHUB-0184: keep the large-play switch reachable; clear the signed-bitwise sweep
GHUB-0184: the sound and large-play switches now lead the toolbar and a
game's actions are appended after them. A QToolBar moves its LAST actions
into the overflow chevron, so right-aligning the two app-wide switches
behind an expanding spacer is what made them the first to be hidden --
Canasta carries eighteen actions and pushed the large-play switch out of
sight at 1400px. Fixed by ordering rather than by the "Play to ..." menu
the item proposed; owner's call, that menu changes a control he uses every
game. Measured with --shot at 1400x760 seed 5: the switch is visible third
from left, and what now sits behind the chevron is Canasta's own last two
entries, Expert partner and Hints. Nothing overflows at 1920.

The expanding spacer went with the reorder. Its job was to push the
switches right; once they lead the bar its only remaining effect was to be
evicted itself, giving a chevron that opened onto an invisible widget.

GHUB-0182: swept src/ under the locally installed clang-tidy (LLVM 23) as
the item asked, rather than fixing the two files it named. 33
bugprone-signed-bitwise findings across NINE files, all fixed; a re-run
reports zero. Three things the item did not predict:

  - the signed operand is often the SHIFT COUNT, not the value, so
    `1u << (d - 1)` still fires and wants `unsigned(d - 1)`;
  - Canasta's `team ^ 1` idiom was six of the sites. teamOf() is
    `seat % 2` over seats 0..3, so `1 - team` is the same value with no
    bitwise operator and no cast;
  - heartsview held a Qt alignment flag in an int, where the fix is the
    correct type rather than a cast. Qt::Alignment then tripped
    bugprone-narrowing-conversions, because drawText takes int flags and
    QFlags converts to unsigned, so that narrowing is now explicit.

The sweep also found 9 performance-use-std-move in four files. Left alone
on the owner's call: they sit on save/restore paths that run once, and are
outside what this item named. So a local run is at zero for
bugprone-signed-bitwise, not at zero outright.

Policy half, settled by the owner 2026-09-06: "a manual run must stay at
zero" means CI's PINNED clang-tidy 18. Check families only grow, so
measuring the rule against whatever happens to be installed locally makes
an LLVM upgrade break it with no code change. Newer local findings are
filed, not treated as breakage.

CLAUDE.md rule 14: docs/specs/GHUB-0017-legibility-switch.md 4.3 stated the
old toolbar mechanism and is amended to match. Recording what was built
rather than changing direction for work still to come, so no gate; that is
rule 14's own amendment instance.

Verified: build clean, ctest 9/9, scripts/local-ci.sh green, clang-tidy
re-run reports zero bugprone-signed-bitwise in src/.

+++ src/chess/chessboard.cpp
@@ -552,7 +552,7 @@ bool Board::insufficientMaterial() const
 {
     int knights = 0;
     int bishops = 0;
-    int bishopSquareColours = 0;   // bit 0 = a light bishop, bit 1 = a dark one
+    unsigned bishopSquareColours = 0;   // bit 0 = a light bishop, bit 1 = a dark one
 
     for (int row = 0; row < kRanks; ++row) {
         for (int col = 0; col < kFiles; ++col) {
@@ -566,7 +566,7 @@ bool Board::insufficientMaterial() const
                 break;
             case PieceType::Bishop:
                 ++bishops;
-                bishopSquareColours |= 1 << ((row + col) % 2);
+                bishopSquareColours |= 1u << unsigned((row + col) % 2);
                 break;
             default:
                 break;
@@ -580,7 +580,7 @@ bool Board::insufficientMaterial() const
         return true;
     // Bishops confined to one colour of square can never deliver mate, however
     // many there are or whoever owns them.
-    return knights == 0 && bishopSquareColours != 3;
+    return knights == 0 && bishopSquareColours != 3u;
 }
 
 std::string Board::notation(const Move& m) const
